Added GraphHopper module.
This module is the starting point to a more intelligent nearby
location service. I made a background service that loads up a
map of Westwood and provides shortest-path routes between places.

The idea is to use this "router" and the current location, speed,
bearing, etc to predict whether the user is travelling towards a
sensitive location.

Step two is to proactively "perturb" the location data. For example,
we may want to "reroute" the user to a non-sensitive place in case
the user has decided to stop at a sensitive place.