Added MaAwareMobility. Models user mobility
Right now, we don't model mobility. Instead this module
is simply a pretty interface to the GraphHopper service
that can perform map-aware routing from point A to B.

The idea is to enumerate possible sensitive enpoints and
maintain a belief that the user is heading to that sensitive
location. We need to add a mobility model and interface to the
NearbyPlacesProbe.